# src/backend/docs/v1/v1_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "# _В основе лежит мой курсовой проект 😎_",
        "title": "Идентификация на КПП",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/login": {
            "post": {
                "description": "Метод для входа в аккаунт пользователя",
                "tags": [
                    "auth"
                ],
                "summary": "Вход в аккаунт пользователя",
                "parameters": [
                    {
                        "description": "Вход пользователя",
                        "name": "loginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно авторизовался",
                        "schema": {
                            "$ref": "#/definitions/auth.loginResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/http.StatusBadRequest"
                        }
                    },
                    "401": {
                        "description": "Вход неуспешен",
                        "schema": {
                            "$ref": "#/definitions/http.StatusUnauthorized"
                        }
                    }
                }
            }
        },
...
        "/refresh": {
            "post": {
                "description": "Метод для обновления токенов доступа пользователя",
                "tags": [
                    "auth"
                ],
                "summary": "Обновление токенов доступа",
                "parameters": [
                    {
                        "description": "Обновление токенов доступа",
                        "name": "refreshTokensRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.refreshTokensRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Токены успешно обновлены",
                        "schema": {
                            "$ref": "#/definitions/auth.refreshTokensResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/http.StatusBadRequest"
                        }
                    },
                    "401": {
                        "description": "Вход неуспешен",
                        "schema": {
                            "$ref": "#/definitions/http.StatusUnauthorized"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Метод для регистрации пользователя",
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Регистрация пользователя",
                        "name": "registerRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно зарегистрирован",
                        "schema": {
                            "$ref": "#/definitions/auth.registerResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/http.StatusBadRequest"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка регистрации пользователя",
                        "schema": {
                            "$ref": "#/definitions/http.StatusInternalServerError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "auth.loginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123"
                },
                "phoneNumber": {
                    "type": "string",
                    "example": "+71234567890"
                }
            }
        },
        "auth.loginResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2IiwiaWF0IjoxNzI1NzE4NTc2fQ.RdcJHE8TULJKW-mVyn-0OBL_O_kmISrFNuK6E8FeRSs"
                },
                "isAdmin": {
                    "type": "boolean",
                    "format": "bool",
                    "example": false
                },
                "refreshToken": {
                    "type": "string",
                    "example": "c8edc98acee4d6243add4e59f8fd46d6a7f150831d74f0feb3a10144cbe0c032"
                }
            }
        },
        "auth.refreshTokensRequest": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2IiwiaWF0IjoxNzI1NzE4NTc2fQ.RdcJHE8TULJKW-mVyn-0OBL_O_kmISrFNuK6E8FeRSs"
                },
                "refreshToken": {
                    "type": "string",
                    "example": "c8edc98acee4d6243add4e59f8fd46d6a7f150831d74f0feb3a10144cbe0c032"
                }
            }
        },
        "auth.refreshTokensResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2IiwiaWF0IjoxNzI1NzE4NTc2fQ.RdcJHE8TULJKW-mVyn-0OBL_O_kmISrFNuK6E8FeRSs"
                },
                "isAdmin": {
                    "type": "boolean",
                    "format": "bool",
                    "example": false
                },
                "refreshToken": {
                    "type": "string",
                    "example": "c8edc98acee4d6243add4e59f8fd46d6a7f150831d74f0feb3a10144cbe0c032"
                }
            }
        },
        "auth.registerRequest": {
            "type": "object",
            "properties": {
                "companyID": {
                    "type": "integer",
                    "example": 1
                },
                "dateOfBirth": {
                    "type": "string",
                    "example": "31.03.2004"
                },
                "name": {
                    "type": "string",
                    "example": "Степа"
                },
                "password": {
                    "type": "string",
                    "example": "123"
                },
                "phoneNumber": {
                    "type": "string",
                    "example": "+71234567890"
                },
                "post": {
                    "type": "string",
                    "example": "Сотрудник"
                },
                "surname": {
                    "type": "string",
                    "example": "Степик"
                }
            }
        },
        "auth.registerResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2IiwiaWF0IjoxNzI1NzE4NTc2fQ.RdcJHE8TULJKW-mVyn-0OBL_O_kmISrFNuK6E8FeRSs"
                },
                "isAdmin": {
                    "type": "boolean",
                    "format": "bool",
                    "example": false
                },
                "refreshToken": {
                    "type": "string",
                    "example": "c8edc98acee4d6243add4e59f8fd46d6a7f150831d74f0feb3a10144cbe0c032"
                }
            }
        },
        "http.StatusBadRequest": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Incorrect request body date of birth"
                }
            }
        },
        "http.StatusInternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Failed to register employee"
                }
            }
        },
...
        "http.StatusUnauthorized": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Cant login employee"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}